Add static file serving and update package.json for build script
The code diff adds static file serving for the 'client/dist' directory and updates the package.json file to include a new 'build' script. The 'index.html' file is now served for all routes. The JSON and URL-encoded middleware are also updated to be used after the static file serving and cookie-parser middleware. The server.js file also includes a description of the project.